+++ SystemML/SystemML/src/main/java/com/ibm/bi/dml/runtime/matrix/data/MatrixValue.java
@@ -106,14 +106,14 @@ public abstract class MatrixValue implements WritableComparable
 
 	public abstract void reset(int rl, int cl, boolean sp);
 	
-	public abstract void reset(int rl, int cl, boolean sp, int nnzs);
+	public abstract void reset(int rl, int cl, boolean sp, long nnzs);
 
 	public abstract void resetDenseWithValue(int rl, int cl, double v) throws DMLRuntimeException ;
 
 	public abstract void copy(MatrixValue that);
 	public abstract void copy(MatrixValue that, boolean sp);
 
-	public abstract int getNonZeros();
+	public abstract long getNonZeros();
 
 	public abstract void setValue(int r, int c, double v);
